Prototype instanced tree generation implemented

+++ Pointspace/Camera.h
@@ -11,7 +11,7 @@
 #include "Shader.h"
 #include "Transform.h"
 
-constexpr int PI = 3.14159265;
+constexpr float PI = 3.14159265;
 
 enum class MovementType {
 	FORWARD,
